requirements.txt
Flask
PyGithub
SQLAlchemy
pymysql